requirements.txt
# 기존 패키지들
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
openai==1.12.0
bleach==6.1.0
redis==5.0.0
httpx==0.25.0
gunicorn==21.2.0
requests==2.31.0

# RAG 시스템용 추가 필요
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.26
chromadb==0.4.22
sentence-transformers==2.2.2
rank-bm25==0.2.2
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.2